requirements.txt
chromadb==0.6.3
coqui_tts==0.24.2
huggingface_hub==0.25.1
langchain==0.3.21
langchain_community==0.3.20
langchain_text_splitters==0.3.7
numpy==2.2.4
PyAudio==0.2.14
pydub==0.25.1
pypdf==5.4.0
python-dotenv==1.1.0
sentence_transformers==3.1.1
sounddevice==0.5.1
torch==2.4.1
TTS==0.22.0
whisperx==3.1.5
